Fix weird lazyredraw issue (amend 2aabf32)
When opening a large file on startup the cursor used to get stuck
at the bottom of the screen and airline wouldn't show until you
made a cursor movement. Apparently this is caused by lazyredraw.
Now lazyredraw is set only after the file is loaded and the cursor
is idle for some time.